upd: updating price, total, installation_fee column type to decimal

+++ database/migrations/2025_01_20_064941_create_packages_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->integer('bandwidth');
             $table->integer('duration');
-            $table->double('price');
+            $table->decimal('price', 12, 2);
             $table->longText('desc');
             $table->timestamps();
         });
